# PYTHON-4144 Optimize json_util performance using single dispatch table