https://jira.mongodb.org/browse/PYTHON-4144

Beanchmark results here: https://spruce.mongodb.com/task/mongo_python_driver_perf_tests_perf_6.0_standalone_patch_cd3f502de6405657c1e825978ec13b295b0d76ef_65a70b9257e85a54e3540958_24_01_16_23_05_08/trend-charts?execution=0

![Screenshot 2024-01-16 at 4 19 18 PM](https://github.com/mongodb/mongo-python-driver/assets/5015933/c89962cf-45ec-49b2-8c46-e95972eb7530)

Results in a 15% improvement to the JSON benchmarks.

Before:
```
$ python -m timeit -s 'from bson.json_util import dumps,loads;doc={"a": list(float(i) for i in range(100000))}' 'dumps(doc)'
10 loops, best of 5: 39 msec per loop
```
After:
```
$ python -m timeit -s 'from bson.json_util import dumps,loads;doc={"a": list(float(i) for i in range(100000))}' 'dumps(doc)'
10 loops, best of 5: 31.1 msec per loop
```