@ShaneHarvey 2024-01-13
I decided to track the perf delta for each BSON type:
<google-sheets-html-origin><style type="text/css"><!--td {border: 1px solid #cccccc;}br {mso-data-placement:same-cell;}--></style>
type | before (secs) | after (secs) | diff %
-- | -- | -- | --
bool | 0.125 | 0.143 | -12.58741259
NoneType | 1.648 | 0.144 | 1044.444444
int | 0.246 | 0.189 | 30.15873016
float | 0.368 | 0.294 | 25.17006803
Int64 | 0.287 | 0.2 | 43.5
str | 0.287 | 0.19 | 51.05263158
bytes | 1.112 | 0.984 | 13.00813008
datetime | 2.816 | 2.699 | 4.334938866
DatetimeMS | 3.878 | 3.696 | 4.924242424
UUID | 1.678 | 1.523 | 10.17728168
Binary | 1.096 | 1.014 | 8.08678501
Code | 0.581 | 0.494 | 17.61133603
DBRef | 3.964 | 3.652 | 8.543263965
ObjectId | 0.592 | 0.506 | 16.99604743
MaxKey | 0.427 | 0.348 | 22.70114943
MinKey | 0.432 | 0.342 | 26.31578947
Regex | 2.935 | 2.947 | -0.4071937564
Pattern | 3.009 | 3.024 | -0.496031746
Timestamp | 0.802 | 0.717 | 11.85495119
Decimal128 | 2.994 | 2.938 | 1.906058543

You can see that this is pretty much an easy win across the board except for `bool` which dropped 12%. The explanation is that in the old code `bool` was the first type we check for so it was being encoded as fast as possible. This advantage (bias) for `bool` disappears now that all type encoding functions are looked up in the same table. Personally I think a small drop for bools is fine given the overall benefit.