@ShaneHarvey 2024-01-13
Note this was the benchmark:
```python
import uuid
import re
import timeit
from datetime import datetime
from bson import *
from bson import json_util

objects = [
    True,
    None,
    1,
    1.1,
    Int64(64),
    'str',
    b'bytes',
    datetime.now(),
    DatetimeMS(1),
    uuid.uuid4(),
    Binary(b'1'),
    Code("code"),
    DBRef("coll", ObjectId()),
    ObjectId(),
    MaxKey(),
    MinKey(),
    Regex("pat"),
    re.compile("pat"),
    Timestamp(1, 1),
    Decimal128("0.5"),
]


for o in objects:
    doc = {"a": [o for _ in range(10000)]}
    opts = json_util.RELAXED_JSON_OPTIONS.with_options(uuid_representation=STANDARD)
    tot = timeit.timeit("json_util.dumps(doc, json_options=opts)", globals=globals(), number=100)
    print(f"{type(o).__name__}, {tot:.3f}")
```